add class and change bit

+++ mypkg/talker.py
@@ -3,17 +3,17 @@
 
 import rclpy
 from rclpy.node import Node
-from std_msgs.msg import Int16
+from std_msgs.msg import Int32
 
 class Talker():
     def __init__(self, node):
-        self.pub = node.create_publisher(Int16, "countup", 10)
+        self.pub = node.create_publisher(Int32, "countup", 10)
         self.fib, self.a, self.b = 0, 0, 1
         self.n = 1
         node.create_timer(1.0, self.cb)
 
     def cb(self):
-        msg = Int16()
+        msg = Int32()
         if self.n == 1:
             msg.data = 1
         else:
